+++ lua/modules/langs/lspconfig.lua
@@ -3,9 +3,6 @@ local lspconfig = require("lspconfig")
 -- if this servers not installed, cmp will callback nvim-lsp-install to install them
 -- fix: how to config according filetype automation install servers
 
--- @see settings.lua
-local settings = require("user.settings")
-
 local capabilities = require("user.capabilities")
 
 local sumneko_lua_locale_adjust = function()
@@ -16,11 +13,11 @@ end
 
 local sumneko_lua_locale = sumneko_lua_locale_adjust()
 
-local navic = require("nvim-navic")
+-- local navic = require("nvim-navic")
 
 local on_attach = function(client, bufnr)
   require("user.lsp_format").on_attach(client)
-  navic.attach(client, bufnr)
+  -- navic.attach(client, bufnr)
 end
 
 -- @ref: https://github.com/lxyoucan/nvim/blob/c84b07f078d20d175a4a3b48a73705b61997bd9f/lua/lspconf/lua.lua#L85
@@ -53,7 +50,7 @@ local lsp_setup = function()
         lspconfig[lsp_server].setup(
           -- settings.lsp_server
           {
-            settings = settings[lsp_server],
+            -- settings = settings[lsp_server],
             on_attach = on_attach,
             capabilities = capabilities,
           }