+++ lua/modules/ui/web_icons.lua
@@ -1,7 +1,19 @@
 require("nvim-web-devicons").setup({
   override = {
+    zsh = {
+      icon = " ",
+      color = "yellow",
+      cterm_color = "67",
+      name = "Zsh",
+    },
+    ["rb"] = {
+      icon = "",
+      color = "#701516",
+      cterm_color = "52",
+      name = "Rb",
+    },
     sh = {
-      icon = "",
+      icon = " ",
       color = "yellow",
       cterm_color = "67",
       name = "Sh",