+++ lua/notify/stages/fade_in_slide_out.lua
@@ -19,19 +19,33 @@ return {
       opacity = 0,
     }
   end,
-  function()
+  function(state, win)
     return {
       opacity = { 100 },
       col = { vim.opt.columns:get() },
+      row = {
+        stages_util.slot_after_previous(win, state.open_windows, stages_util.DIRECTION.TOP_DOWN),
+        frequency = 3,
+        complete = function()
+          return true
+        end,
+      },
     }
   end,
-  function()
+  function(state, win)
     return {
       col = { vim.opt.columns:get() },
       time = true,
+      row = {
+        stages_util.slot_after_previous(win, state.open_windows, stages_util.DIRECTION.TOP_DOWN),
+        frequency = 3,
+        complete = function()
+          return true
+        end,
+      },
     }
   end,
-  function()
+  function(state, win)
     return {
       width = {
         1,
@@ -49,6 +63,13 @@ return {
         end,
       },
       col = { vim.opt.columns:get() },
+      row = {
+        stages_util.slot_after_previous(win, state.open_windows, stages_util.DIRECTION.TOP_DOWN),
+        frequency = 3,
+        complete = function()
+          return true
+        end,
+      },
     }
   end,
 }