+++ lua/user/capabilities.lua
@@ -25,6 +25,7 @@ vim.lsp.handlers["textDocument/signatureHelp"] = vim.lsp.with(vim.lsp.handlers.s
 
 -- for null-ls bug
 capabilities.offsetEncoding = { "utf-16" }
+-- @ref: https://github.com/NvChad/NvChad/commit/429ce94da17d850ea022cf7bf188ec676e4c7be7
 -- capabilities.documentFormattingProvider = false
 -- capabilities.documentRangeFormattingProvider = false